email is now a mandatory field on GET for persist.  I've also restricted the data that is returned to the bare minimum (id, email, stripe authCode, promoCode and dateAdded).  If you omit the email parameter you will be told that it is mandatory.
I've also modified the POST method so that if you include the id (which is returned from the POST method) it will remove that node, and save the new data, returning a new id each time - there doesn't seem to be an update API method.  If you use something like:

{"id": "017d31e0-70e0-46dd-a7d0-f66cbb2bd6eb","scopeUser": {"email": "frank@franky.com","firstName": "Ferenc","city": "London"},"promoCode": "INSTANT","authCode": "ZAQ123","dateAdded": "Sat Apr 22 21:08:02 BST 2017"}

as POST data - notice the "id" attribute in there, it will remove the existing node, and save the new data.